refactor: enforce single component per file and tune oxlint rules
- 複数コンポーネントを持つファイルをリポジトリ規約(1ファイル1コンポーネント)に従い kebab-case ファイルへ分割し、react(no-multi-comp) の警告25件を解消
- ui-state.tsx を ui-loading / ui-empty / ui-error へ分割
- 重複していた Error fallback を createErrorFallback ファクトリに集約
- oxlint.config.ts: 教条的なルール(no-null, no-undefined, no-ternary, no-optional-chaining, jsx-max-depth, jsx-no-literals, forbid-component-props)を無効化
- src/routes/** は TanStack Router 規約のため no-multi-comp を override、styled-link / stories catalog は理由付きでファイル単位無効化
