+++ src/features/app-shell/components/shelf-nav-panel.tsx
@@ -1,21 +1,14 @@
 import { useRouter } from '@tanstack/react-router'
 import { Suspense } from 'react'
-import { ErrorBoundary, getErrorMessage } from 'react-error-boundary'
-import type { FallbackProps } from 'react-error-boundary'
+import { ErrorBoundary } from 'react-error-boundary'
 
-import { UiError, UiLoading } from '../../../shared/components/ui-state'
+import { createErrorFallback } from '../../../shared/components/error-fallback'
+import { UiLoading } from '../../../shared/components/ui-loading'
 import type { ShelfNavSelection } from '../../tags/components/shelf-nav'
-import { ShelfNavAsync } from '../../tags/components/shelf-nav'
+import { ShelfNavAsync } from '../../tags/components/shelf-nav-async'
 import type { ShelfTag } from '../../tags/lib/tag-shelf'
 
-function ShelfNavError({ error, resetErrorBoundary }: FallbackProps) {
-  return (
-    <UiError
-      message={getErrorMessage(error) ?? '棚の読み込みに失敗しました'}
-      onRetry={resetErrorBoundary}
-    />
-  )
-}
+const ShelfNavError = createErrorFallback('棚の読み込みに失敗しました')
 
 export function ShelfNavPanel({
   shelfTagsPromise,
